Add purge frequency to web/UserSyncHandler.TidyUp
Previously everytime a DB file is opened a purge will be
attempted. Users with a large number of BSOs will incur a full table
scan on BSO.ttl which causes a lot of IO.

The prototype server also shows that out of 66,309 purges only 67 (0.1%)
resulted in any actual deletion of data. This is much too aggressive.

Changes:
- TidyUp will skip purging if not enough time has past since last tidy
- Pool will only cleanup when a db is opened, once every 12 hours